% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlBLM_Transform.R
\name{AlBLM}
\alias{AlBLM}
\title{Aluminum BLM}
\usage{
AlBLM(x)
}
\arguments{
\item{x}{table output from AWQMS query}
}
\value{
Dataframe with input file for BLM
}
\description{
A function that converts data from AWQMS long table format into the wide table format that
is used by the Aluminum BLM model and the NPDES Aluminum BLM templates. Ancillary data is
the first data found in the calendar day.
}
\details{
Output includes the method detection and method reporting limit for Aluminum, as well as result
value and also the "Result_Type" for each characteristic so that
user can better judge the quality of the data. This function does not return data that has
a "Rejected" status in AWQMS.

Written by Aliana Britson
}
