% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Bio_Indexes.R
\name{AWQMS_Bio_Indexes}
\alias{AWQMS_Bio_Indexes}
\title{AWQMS_Bio_Indexes}
\usage{
AWQMS_Bio_Indexes(
  startdate = NULL,
  enddate = NULL,
  MLocID = NULL,
  AU_ID = NULL,
  HUC12_Name = NULL,
  project = NULL,
  OrganizationID = NULL,
  ReferenceSite = NULL,
  Index_Name = NULL,
  DQL = NULL,
  return_query = FALSE
)
}
\arguments{
\item{startdate}{Required parameter setting the startdate of the data being fetched. Format 'yyyy-mm-dd'. Defaults to
'1949-09-15'}

\item{enddate}{Optional parameter setting the enddate of the data being fetched. Format 'yyyy-mm-dd'}

\item{AU_ID}{Optional vector of Assessment Units to be fetched}

\item{project}{Optional vector of projects to be fetched}

\item{ReferenceSite}{filter for reference or non reference streams. NOT CURRENTLY POPULATED}

\item{Index_Name}{Optional vector of metrics to filter}

\item{DQL}{Optional vector of DQLs to fetch}

\item{return_query}{If FALSE, fetches data from AWQMS. If TRUE, returns string of query language.}

\item{station}{Optional vector of stations to be fetched}

\item{org}{optional vector of Organizations to be fetched}
}
\description{
Thus function will retrieve Bio_Indexes data from Oregon DEQ AWQMS
}
