% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Bio_Metrics.R
\name{AWQMS_Bio_Metrics}
\alias{AWQMS_Bio_Metrics}
\title{AWQMS_Bio_Metrics}
\usage{
AWQMS_Bio_Metrics(
  startdate = "1949-09-15",
  enddate = NULL,
  station = NULL,
  AU_ID = NULL,
  HUC12_Name = NULL,
  project = NULL,
  org = NULL,
  ReferenceSite = NULL,
  Metric_Name = NULL,
  DQL = NULL,
  return_query = FALSE
)
}
\arguments{
\item{startdate}{Required parameter setting the startdate of the data being fetched. Format 'yyyy-mm-dd'. Defaults to
'1949-09-15'}

\item{enddate}{Optional parameter setting the enddate of the data being fetched. Format 'yyyy-mm-dd'}

\item{station}{Optional vector of stations to be fetched}

\item{AU_ID}{Optional vector of Assessment Units to be fetched}

\item{project}{Optional vector of projects to be fetched}

\item{org}{optional vector of Organizations to be fetched}

\item{ReferenceSite}{filter for reference or non reference streams. NOT CURRENTLY POPULATED}

\item{Metric_Name}{Optional vector of metrics to filter}

\item{DQL}{Optional vector of DQLs to fetch}

\item{return_query}{If FALSE, fetches data from AWQMS. If TRUE, returns string of query language.}
}
\description{
Thus function will retrieve Bio_Metrics data from Oregon DEQ AWQMS
}
