% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Chars.R
\name{AWQMS_Chars}
\alias{AWQMS_Chars}
\title{AWQMS_Chars}
\usage{
AWQMS_Chars(project = NULL, station = NULL)
}
\arguments{
\item{project}{Optional vector of projects to be filtered on}

\item{station}{Optional vector of stations to be filtered on}
}
\value{
Dataframe of monitoring locations
}
\description{
This function will return a list of characterustucs found in OregonDEQ AWQMS
}
\examples{
AWQMS_Chars(project = 'Total Maximum Daily Load Sampling', station = c('29820-ORDEQ', '10724-ORDEQ'))
}
