% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Data_Cont.R
\name{AWQMS_Data_Cont}
\alias{AWQMS_Data_Cont}
\title{AWQMS_Data_Cont}
\usage{
AWQMS_Data_Cont(
  startdate = "1949-09-15",
  enddate = NULL,
  station = NULL,
  AU_ID = NULL,
  char = NULL,
  media = NULL,
  org = NULL,
  HUC8 = NULL,
  HUC8_Name = NULL,
  HUC10 = NULL,
  HUC12 = NULL,
  HUC12_Name = NULL,
  Result_Status = NULL,
  crit_codes = FALSE
)
}
\arguments{
\item{startdate}{Required parameter setting the startdate of the data being fetched. Format 'yyyy-mm-dd'}

\item{enddate}{Optional parameter setting the enddate of the data being fetched. Format 'yyyy-mm-dd'}

\item{station}{Optional vector of stations to be fetched}

\item{AU_ID}{Optional vector of Assessment units to be fetched}

\item{char}{Optional vector of characteristics to be fetched}

\item{media}{Optional vector of sample media to be fetched}

\item{org}{optional vector of Organizations to be fetched}

\item{HUC8}{Optional vector of HUC8s to be fetched}

\item{HUC8_Name}{Optional vector of HUC8 names to be fetched}

\item{HUC10}{Optional vector of HUC10s to be fetched}

\item{HUC12}{Optional vector of HUC12s to be fetched}

\item{HUC12_Name}{Optional vector of HUC12 names to be fetched}

\item{Result_Status}{oprtion vector of result status to be fetched. Set to Result_Status = 'Final' if only data that has passed QAQC is included}

\item{crit_codes}{If true, include standard codes used in determining criteria}
}
\value{
Dataframe of data from AWQMS
}
\description{
This function will retrive raw continuous data from OregonDEQ AWQMS
}
\examples{
AWQMS_Data_Cont(station = '39446-ORDEQ', Result_Status = 'Final')
}
