% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Orgs.R
\name{AWQMS_Orgs}
\alias{AWQMS_Orgs}
\title{AWQMS_ORGS}
\usage{
AWQMS_Orgs(project = NULL, station = NULL)
}
\arguments{
\item{project}{Optional vector of projects to be filtered on}

\item{station}{Optional vector of stations to be filtered on}
}
\value{
Dataframe of organizations with available data
}
\description{
This function will return a list of organizations with data in AWQMS found in OregonDEQ AWQMS
}
\examples{
AWQMS_Orgs(project = 'Total Maximum Daily Load Sampling', station = c('29820-ORDEQ', '10724-ORDEQ'))
}
