% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Projects.R
\name{AWQMS_Projects}
\alias{AWQMS_Projects}
\title{AWQMS_Projects}
\usage{
AWQMS_Projects()
}
\value{
Dataframe of projects in AWQMS
}
\description{
This function will create a vector or projects from AWQMS Project1, project2, and project3 columns.
}
\examples{
AWQMS_Projects()


}
