% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Raw_Macros.R
\name{AWQMS_Raw_Macros}
\alias{AWQMS_Raw_Macros}
\title{AWQMS_Raw_Macros}
\usage{
AWQMS_Raw_Macros(
  startdate = "1949-09-15",
  enddate = NULL,
  station = NULL,
  AU_ID = NULL,
  project = NULL,
  org = NULL,
  Wade_Boat = NULL,
  ReferenceSite = NULL,
  HUC8 = NULL,
  HUC8_Name = NULL,
  HUC10 = NULL,
  HUC12 = NULL,
  HUC12_Name = NULL,
  Char_Name = NULL,
  Bio_Intent = NULL,
  Taxonomic_Name = NULL,
  StageID = NULL,
  UniqueTaxon = NULL,
  Analytical_method = NULL,
  return_query = FALSE
)
}
\arguments{
\item{startdate}{Required parameter setting the startdate of the data being fetched. Format 'yyyy-mm-dd'. Defaults to
'1949-09-15'}

\item{enddate}{Optional parameter setting the enddate of the data being fetched. Format 'yyyy-mm-dd'}

\item{station}{Optional vector of stations to be fetched}

\item{AU_ID}{Optional vector of Assessment Units to be fetched}

\item{project}{Optional vector of projects to be fetched}

\item{org}{optional vector of Organizations to be fetched}

\item{Wade_Boat}{option for filtering wadeable or boat samples. Options are: 'wadeable', 'boatable', or NULL.}

\item{ReferenceSite}{filter for reference or non reference streams. Options are: NULL, '<Null>', 'YES', 'Not Loaded',
"".}

\item{HUC8}{Optional vector of HUC8s to be fetched}

\item{HUC8_Name}{Optional vector of HUC8 names to be fetched}

\item{HUC10}{Optional vector of HUC10s to be fetched}

\item{HUC12}{Optional vector of HUC12s to be fetched}

\item{HUC12_Name}{Optional vector of HUC12 names to be fetched}

\item{Char_Name}{Optional vector or characteristic names. Options are: 'Count', 'Density'}

\item{Bio_Intent}{Optional vector or Bio_intent. Options are: 'Population Census', 'Species Density'}

\item{Taxonomic_Name}{Optional vector of taxa}

\item{StageID}{Optional vector of stages. Options are: NULL, 'Adult', 'Larva', 'Pupa'}

\item{UniqueTaxon}{Optioal vector. Options are: 'UniqueTaxon', 'AmbiguousTaxon'}

\item{return_query}{If FALSE, fetches data from AWQMS. If TRUE, returns string of query language.}
}
\value{
Dataframe of data from AWQMS
}
\description{
Thus function will retrieve raw macroinvertebrate data from Oregon DEQ AWQMS
}
