% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Stations.R
\name{AWQMS_Station}
\alias{AWQMS_Station}
\title{AWQMS_Station}
\usage{
AWQMS_Station(project = NULL, char = NULL, HUC8 = NULL)
}
\arguments{
\item{project}{Optional vector of projects to be filtered on}

\item{HUC8}{Optional vector of HUC8s to be filtered on}

\item{station}{Optional vector of stations to be filtered on}
}
\value{
Dataframe of monitoring locations
}
\description{
This function will return a list of monitoring locations with data found in OregonDEQ AWQMS
}
\examples{
AWQMS_Station(project = 'Total Maximum Daily Load Sampling', char = "Temperature, water", HUC8 = "17090003")
}
