% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Stations.R
\name{AWQMS_Stations}
\alias{AWQMS_Stations}
\title{AWQMS_Station}
\usage{
AWQMS_Stations(
  project = NULL,
  char = NULL,
  HUC8 = NULL,
  HUC8_Name = NULL,
  org = NULL,
  crit_codes = FALSE
)
}
\arguments{
\item{project}{Optional vector of projects to be filtered on}

\item{char}{Optional vector of characters to be filtered on}

\item{HUC8}{Optional vector of HUC8s to be filtered on}

\item{HUC8_Name}{Optional vector of HUC8 names to be filtered on}

\item{org}{Optional vector of organizations to be filtered on}

\item{crit_codes}{If true, include standard codes used in determining criteria}
}
\value{
Dataframe of monitoring locations
}
\description{
This function will return a list of monitoring locations with data found in OregonDEQ AWQMS
}
\examples{
AWQMS_Station(project = 'Total Maximum Daily Load Sampling', char = "Temperature, water", HUC8 = "17090003", crit_codes = false)
}
