% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Stations_strds.R
\name{AWQMS_Stations_strds}
\alias{AWQMS_Stations_strds}
\title{AWQMS_Station_strds}
\usage{
AWQMS_Stations_strds(project = NULL, char = NULL, HUC8 = NULL,
  HUC8_Name = NULL, org = NULL)
}
\arguments{
\item{project}{Optional vector of projects to be filtered on}

\item{char}{Optional vector of characters to be filtered on}

\item{HUC8}{Optional vector of HUC8s to be filtered on}

\item{HUC8_Name}{Optional vector of HUC8 names to be filtered on}

\item{org}{Optional vector of organizations to be filtered on}
}
\value{
Dataframe of monitoring locations
}
\description{
This function will return a list of monitoring locations with data found in OregonDEQ AWQMS
Joined witgh the standard codes in the Stations Database
}
\examples{
AWQMS_Stations_strds(project = 'Total Maximum Daily Load Sampling', char = "Temperature, water", HUC8 = "17090003")
}
