% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bact_crit}
\alias{Bact_crit}
\title{Bacteria criteria}
\format{
A dataframe
}
\usage{
Bact_crit
}
\description{
An obsolete dataframe of the Bacteria criteria. Kept for legacy purposes
}
\keyword{datasets}
