% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LU_BacteriaCode}
\alias{LU_BacteriaCode}
\title{Bacteria Code Lookup Table}
\format{
A dataframe
}
\usage{
LU_BacteriaCode
}
\description{
A dataframe to lookup Bacteria Code defnitions
}
\keyword{datasets}
