% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LU_BenUseCode}
\alias{LU_BenUseCode}
\title{Benefical uses lookup table}
\format{A dataframe}
\usage{
LU_BenUseCode
}
\description{
A dataframe to lookup beneficial uses by ben_use_code
}
\keyword{datasets}
