% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LU_DOCode}
\alias{LU_DOCode}
\title{Dissolved Oxygen Code Lookup Table}
\format{
A dataframe
}
\usage{
LU_DOCode
}
\description{
A dataframe to lookup dissolved oxygen code defnitions
}
\keyword{datasets}
