% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_multiorg_sites.R
\name{missing_multiorg_sites}
\alias{missing_multiorg_sites}
\title{missing_multiorg_sites}
\usage{
missing_multiorg_sites(save_location)
}
\arguments{
\item{save_location}{}
}
\value{
writes an excel file
}
\description{
This function will find monitoring locations that have multiple organizations
associated with the same MLocID. MLocIDs with more than 1 org associated with
it are passed to the monitoring_locations_vw view in the stations database to
see if they exist. Join failures are then written to an excel file at the
save_location identified in the function to be added to the
monitoring_locations_vw view in the stations database
}
