% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mlocs_crit.R
\name{Mlocs_crit}
\alias{Mlocs_crit}
\title{Query ODEQ's Stations database and return a table of siote specific criteria}
\usage{
Mlocs_crit(mlocs = NULL, stations_odbc = "STATIONS")
}
\arguments{
\item{mlocs}{Vector of unique monitoring location station IDs (MLocIDs).
#' @keywords stations, criteria}

\item{stations_odbc}{Stations database ODBC system data source name (DSN) identifed in the ODBC data sources administrator. Default is "STATIONS".}
}
\value{
Dataframe from the stations database joined with site specific criteria
}
\description{
Retrieve station information from ODEQ's Stations database and joines with site specific criteria tables
}
\details{
A vector of monitoring locations must be supplies.
This function will only work for employees of ODEQ. Requires read access permissions for internal odbc connections to
the AWQMS and Stations databases.
}
\examples{
library(AWQMSdata)

}
