% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ToxHH_crit}
\alias{ToxHH_crit}
\title{Human Health toxic criteria}
\format{
A dataframe
}
\usage{
ToxHH_crit
}
\description{
A dataframe of the  Human Health toxic criteria
}
\keyword{datasets}
