% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Data.R
\name{AWQMS_Data}
\alias{AWQMS_Data}
\title{AWQMS_Data}
\usage{
AWQMS_Data(startdate = "1949-09-15", enddate = NULL, station = NULL,
  project = NULL, char = NULL, stat_base = NULL, media = NULL,
  org = NULL, HUC8 = NULL, HUC8_Name = NULL, HUC10 = NULL,
  HUC12 = NULL, HUC12_Name = NULL, crit_codes = FALSE,
  filterQC = TRUE)
}
\arguments{
\item{startdate}{Required parameter setting the startdate of the data being fetched. Format 'yyyy-mm-dd'}

\item{enddate}{Optional parameter setting the enddate of the data being fetched. Format 'yyyy-mm-dd'}

\item{station}{Optional vector of stations to be fetched}

\item{project}{Optional vector of projects to be fetched}

\item{char}{Optional vector of characteristics to be fetched}

\item{stat_base}{Optional vector of Result Stattistical Bases to be fetched ex. Maximum}

\item{media}{Optional vector of sample media to be fetched}

\item{org}{optional vector of Organizations to be fetched}

\item{HUC8}{Optional vector of HUC8s to be fetched}

\item{HUC8_Name}{Optional vector of HUC8 names to be fetched}

\item{HUC10}{Optional vector of HUC10s to be fetched}

\item{HUC12}{Optional vector of HUC12s to be fetched}

\item{HUC12_Name}{Optional vector of HUC12 names to be fetched}

\item{crit_codes}{If true, include standard codes used in determining criteria}

\item{filterQC}{If true, do not return MLocID 10000-ORDEQ or sample replicates}
}
\value{
Dataframe of data from AWQMS
}
\description{
This function will retrive data from OregonDEQ AWQMS
}
\examples{
AWQMS_Data(startdate = '2017-1-1', enddate = '2000-12-31', station = c('10591-ORDEQ', '29542-ORDEQ'),
project = 'Total Maximum Daily Load Sampling', filterQC = FALSE, crit_codes = FALSE)
}
