% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AWQMS_Data.R
\name{AWQMS_Data}
\alias{AWQMS_Data}
\title{AWQMS_Data}
\usage{
AWQMS_Data(
  startdate = "1949-09-15",
  enddate = NULL,
  MLocID = NULL,
  AU_ID = NULL,
  project = NULL,
  Char_Name = NULL,
  CASNumber = NULL,
  Statistical_Base = NULL,
  SampleMedia = NULL,
  SampleSubmedia = NULL,
  OrganizationID = NULL,
  HUC8 = NULL,
  HUC8_Name = NULL,
  HUC10 = NULL,
  HUC12 = NULL,
  HUC12_Name = NULL,
  EcoRegion3 = NULL,
  crit_codes = FALSE,
  filterQC = TRUE,
  return_query = FALSE
)
}
\arguments{
\item{startdate}{Optional parameter setting the startdate of the data being fetched. Format 'yyyy-mm-dd'}

\item{enddate}{Optional parameter setting the enddate of the data being fetched. Format 'yyyy-mm-dd'}

\item{MLocID}{Optional vector of stations to be fetched}

\item{AU_ID}{Optional vector of Assessment Units to be fetched}

\item{project}{Optional vector of projects to be fetched}

\item{Char_Name}{Optional vector of characteristics to be fetched}

\item{CASNumber}{Optional vector of CAS numbers to be fetched}

\item{Statistical_Base}{Optional vector of Result Stattistical Bases to be fetched ex. Maximum}

\item{SampleMedia}{Optional vector of sample media to be fetched}

\item{SampleSubmedia}{Optional vector of sample submedia to be fetched}

\item{OrganizationID}{optional vector of Organizations to be fetched}

\item{HUC8}{Optional vector of HUC8s to be fetched}

\item{HUC8_Name}{Optional vector of HUC8 names to be fetched}

\item{HUC10}{Optional vector of HUC10s to be fetched}

\item{HUC12}{Optional vector of HUC12s to be fetched}

\item{HUC12_Name}{Optional vector of HUC12 names to be fetched}

\item{crit_codes}{If true, include standard codes used in determining criteria}

\item{filterQC}{If true, do not return MLocID 10000-ORDEQ or sample replicates}

\item{return_query}{If true, return the query language that would have been sent to AWQMS. If TRUE, nothing is sent to AWQMS. This is useful for troubleshooting.}
}
\value{
Dataframe of data from AWQMS
}
\description{
This function will retrieve data from OregonDEQ AWQMS
}
\examples{
AWQMS_Data(startdate = '2000-1-1', enddate = '2000-12-31', MLocID = c('10591-ORDEQ', '29542-ORDEQ'),
project = 'Total Maximum Daily Load Sampling', filterQC = FALSE, crit_codes = FALSE)
}
