% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CuBLM_Transform.R
\name{CuBLM}
\alias{CuBLM}
\title{CuBLM}
\usage{
CuBLM(x)
}
\arguments{
\item{x}{table output from AWQMS query}
}
\value{
Dataframe with input file for BLM
}
\description{
A function that converts data from AWQMS long table format into the wide table format that
res saying that multiple rows match, look for duplicates in data. Written by Aliana Britson
}
